# Add socket tree and schemas caching